refactor: abstract user role perms check

+++ apps/backend/src/app/modules/role/role.controller.ts
@@ -19,9 +19,10 @@ import type {
 } from "@shared/common/types";
 import { RoleInterceptor, Roles } from "@interceptors";
 import { AppError, AppErrorTypes } from "@utils/appErrors";
-import { hasPermission } from "@shared/common/abac";
+import { rolesHavePermission } from "@utils/rolesHavePermission";
 
 @Controller("role")
+@UseInterceptors(RoleInterceptor)
 export class RoleController {
 	constructor(
 		private readonly roleService: RoleService,
@@ -29,16 +30,11 @@ export class RoleController {
 	) {}
 
 	@Post()
-	@UseInterceptors(RoleInterceptor)
 	async create(
 		@Body(new ValidationPipe(CreateRoleInputSchema)) data: CreateRoleInput,
 		@Roles() roles: StrictRole[],
 	) {
-		if (
-			roles.some((role) =>
-				role.policies.some((policy) => hasPermission(policy, "role", "readWrite")),
-			)
-		) {
+		if (rolesHavePermission(roles, "role", "readWrite")) {
 			return await this.roleService.create(data);
 		}
 		throw new AppError(AppErrorTypes.NoPermission);
@@ -53,108 +49,82 @@ export class RoleController {
 	}
 
 	@Delete(":id")
-	async deleteById(@Param("id", new ValidationPipe(GetRoleInputSchema)) id: string) {
-		return await this.roleService.deleteById(id);
+	async deleteById(
+		@Param("id", new ValidationPipe(GetRoleInputSchema)) id: string,
+		@Roles() roles: StrictRole[],
+	) {
+		if (rolesHavePermission(roles, "role", "readWrite", { id })) {
+			return await this.roleService.deleteById(id);
+		}
 	}
 
 	@Get("all")
-	@UseInterceptors(RoleInterceptor)
 	async getAll(@Roles() roles: StrictRole[]) {
-		if (
-			roles.some((role) => role.policies.some((policy) => hasPermission(policy, "role", "read")))
-		) {
+		if (rolesHavePermission(roles, "role", "read")) {
 			return await this.roleService.getAll();
 		}
 		throw new AppError(AppErrorTypes.NoPermission);
 	}
 
 	@Get(":id")
-	@UseInterceptors(RoleInterceptor)
 	async getById(
 		@Param("id", new ValidationPipe(GetRoleInputSchema)) id: string,
 		@Roles() roles: StrictRole[],
 	) {
-		if (
-			roles.some((role) =>
-				role.policies.some((policy) => hasPermission(policy, "role", "read", { id })),
-			)
-		) {
+		if (rolesHavePermission(roles, "role", "read", { id })) {
 			return await this.roleService.getById(id);
 		}
 		throw new AppError(AppErrorTypes.NoPermission);
 	}
 
 	@Get("policy/all")
-	@UseInterceptors(RoleInterceptor)
 	async getAllPolicies(@Roles() roles: StrictRole[]) {
-		if (
-			roles.some((role) => role.policies.some((policy) => hasPermission(policy, "policy", "read")))
-		) {
+		if (rolesHavePermission(roles, "policy", "read")) {
 			return await this.policyService.getAll();
 		}
 		throw new AppError(AppErrorTypes.NoPermission);
 	}
 
 	@Get("policy/:id")
-	@UseInterceptors(RoleInterceptor)
 	async getPolicyById(
 		@Param("id", new ValidationPipe(GetPolicyInputSchema)) id: string,
 		@Roles() roles: StrictRole[],
 	) {
-		if (
-			roles.some((role) =>
-				role.policies.some((policy) => hasPermission(policy, "policy", "read", { id })),
-			)
-		) {
+		if (rolesHavePermission(roles, "policy", "read", { id })) {
 			return await this.policyService.getById(id);
 		}
 		throw new AppError(AppErrorTypes.NoPermission);
 	}
 
 	@Post("policy")
-	@UseInterceptors(RoleInterceptor)
 	async createPolicy(
 		@Body(new ValidationPipe(CreatePolicyInputSchema)) data: CreatePolicyInput,
 		@Roles() roles: StrictRole[],
 	) {
-		if (
-			roles.some((role) =>
-				role.policies.some((policy) => hasPermission(policy, "policy", "readWrite")),
-			)
-		) {
+		if (rolesHavePermission(roles, "policy", "readWrite")) {
 			return await this.policyService.create(data);
 		}
 		throw new AppError(AppErrorTypes.NoPermission);
 	}
 
 	@Patch("policy/:id")
-	@UseInterceptors(RoleInterceptor)
 	async updatePolicyById(
 		@Param("id", new ValidationPipe(GetPolicyInputSchema)) id: string,
 		@Body(new ValidationPipe(UpdatePolicyInputSchema)) data: UpdatePolicyInput,
 		@Roles() roles: StrictRole[],
 	) {
-		if (
-			roles.some((role) =>
-				role.policies.some((policy) => hasPermission(policy, "policy", "readWrite", { id })),
-			)
-		) {
+		if (rolesHavePermission(roles, "policy", "readWrite", { id })) {
 			return await this.policyService.updateById(id, data);
 		}
 		throw new AppError(AppErrorTypes.NoPermission);
 	}
 
 	@Delete("policy/:id")
-	@UseInterceptors(RoleInterceptor)
 	async deletePolicyById(
 		@Param("id", new ValidationPipe(GetPolicyInputSchema)) id: string,
 		@Roles() roles: StrictRole[],
 	) {
-		if (
-			roles.some((role) =>
-				role.policies.some((policy) => hasPermission(policy, "policy", "readWrite", { id })),
-			)
-		) {
+		if (rolesHavePermission(roles, "policy", "readWrite", { id })) {
 			return await this.policyService.deleteById(id);
 		}
 		throw new AppError(AppErrorTypes.NoPermission);
